requirements.txt
requests==2.31.0
selenium==4.15.2
playwright==1.40.0
PyPDF2==3.0.1
pandas==2.1.4
python-dotenv==1.0.0
schedule==1.2.0
openpyxl==3.1.2
beautifulsoup4==4.12.2
lxml==4.9.3
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9
